seperate out ui logic
